+++ src/options-conflict-resolver/base.js
@@ -1,6 +1,0 @@
-export default class OptionsConflictResolver {
-    // eslint-disable-next-line no-unused-vars
-    predicate( widget, options ) {}
-    // eslint-disable-next-line no-unused-vars
-    resolve( options ) {}
-}